requirements.txt
telethon==1.26.1
swibots
cryptg
python-decouple